Use Utf8 coding instead of Latin1 when handling journald messages
Assume UTF-8 coding instead of Latin 1 for strings when parsing from and
passing to the journal.

The sd_journal_print() documentation recommends applications use UTF-8
when logging. It is likely that applications do log in UTF-8, if only
because its a typical default in locale set up (e.g. US_EN.UTF8).
BUG: 350070